+++ src/navigation/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {createStackNavigator} from '@react-navigation/stack';
 import {NavigationContainer} from '@react-navigation/native';
-import Home from '../screens/Home/Home';
+import HomeScreen from '../screens/Home/HomeScreen';
 import LearningScreen from '../screens/LearningScreen/LearningScreen';
 
 const Stack = createStackNavigator();
@@ -10,7 +10,7 @@ const CurrentListStack = () => {
   return (
     <NavigationContainer>
       <Stack.Navigator screenOptions={{headerShown: false}}>
-        <Stack.Screen name="Home" component={Home} />
+        <Stack.Screen name="HomeScreen" component={HomeScreen} />
         <Stack.Screen name="LearningScreen" component={LearningScreen} />
       </Stack.Navigator>
     </NavigationContainer>
